requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.1
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
scikit-fuzzy>=0.4.2
numpy>=1.26.4
scipy>=1.13.0
umap-learn>=0.5.6
matplotlib>=3.8.4
seaborn>=0.13.2
pandas>=2.2.2
tqdm>=4.66.4
python-dotenv>=1.0.1